+++ src/material.cpp
@@ -6,37 +6,37 @@
 using namespace glm;
 using namespace std;
 
-vec3 MaterialUniform::shade(Intersection, Scene&)
+dvec3 MaterialUniform::shade(Intersection, Scene&)
 {
     return color;
 }
 
-vec3 MaterialLambert::shade(Intersection inter, Scene& scene)
+dvec3 MaterialLambert::shade(Intersection inter, Scene& scene)
 {
     // À COMPLÉTER
-    return vec3(0);
+    return dvec3(0);
 }
 
-vec3 MaterialLambertTextured::shade(Intersection inter, Scene& scene)
+dvec3 MaterialLambertTextured::shade(Intersection inter, Scene& scene)
 {
     // À COMPLÉTER
-    return vec3(0);
+    return dvec3(0);
 }
 
-vec3 MaterialPhong::shade(Intersection inter, Scene& scene)
+dvec3 MaterialPhong::shade(Intersection inter, Scene& scene)
 {
     // À COMPLÉTER
-    return vec3(0);
+    return dvec3(0);
 }
 
-vec3 MaterialReflective::shade(Intersection inter, Scene &scene)
+dvec3 MaterialReflective::shade(Intersection inter, Scene &scene)
 {
     // À COMPLÉTER
-    return vec3(0);
+    return dvec3(0);
 }
 
-vec3 MaterialRefractive::shade(Intersection inter, Scene &scene)
+dvec3 MaterialRefractive::shade(Intersection inter, Scene &scene)
 {
     // À COMPLÉTER
-    return vec3(0);
+    return dvec3(0);
 }
