+++ src/texture.h
@@ -11,18 +11,18 @@ public:
     Texture() : name() {}
 
     // retourne la couleur de la texture pour des coordonnées UV données.
-    virtual glm::vec3 getColor(glm::vec2 uv) = 0;
+    virtual glm::dvec3 getColor(glm::dvec2 uv) = 0;
 };
 
 class TextureCheckerboard : public Texture
 {
 public:
     int nbSquaresX, nbSquaresY;
-    glm::vec3 color1, color2;
+    glm::dvec3 color1, color2;
 
     TextureCheckerboard() :
         nbSquaresX(16), nbSquaresY(16), color1(0), color2(1) {}
-    glm::vec3 getColor(glm::vec2 uv);
+    glm::dvec3 getColor(glm::dvec2 uv);
 };
 
 class TextureImageFile : public Texture
@@ -32,7 +32,7 @@ public:
     bool bilinear;
 
     TextureImageFile(std::string filename) : image(filename), bilinear(false) {}
-    glm::vec3 getColor(glm::vec2 uv);
+    glm::dvec3 getColor(glm::dvec2 uv);
     unsigned int width() {return image.getNbPixelsX();}
     unsigned int height() {return image.getNbPixelsY();}
 };
